Clean test case naming
